#1 decouple initialize method and expose main.
In the context of webpack, it is hard to access `document` without workarounds. In addition, the main method that really needs to be exposed is the `shadowStyles` method. The rest of the methods and functions could even become dependencies and imported into the project.

In addition, because `shadowstyles` is dependent on MutationObserver, it make sense to require it in the beginning. That is, if the context is commonJS then use the polyfill installed from `npm`, otherwise use the `MutationObserver` that exists on `window` that could have come from either a pollyfill defined on `window` or just defined natively by the browser.

Signed-off-by: Amin Meyghani <meyghania@gmail.com>